+++ tdmclient/tools/__init__.py
@@ -1,4 +1,9 @@
 """Tools to be launched from a terminal or from outside Python.
 """
 
-from . import (gui, list, repl, run, server, tdmdiscovery, transpile, watch)
+from . import (list, repl, run, server, tdmdiscovery, transpile, watch)
+try:
+    # weak import in case tkinter isn't available
+    from . import gui
+except ModuleNotFoundError:
+    pass